Another batch of printf format fixes.
Original commit message from CVS:
* ext/dts/gstdtsdec.c: (gst_dtsdec_chain):
* ext/musicbrainz/gsttrm.c: (gst_trm_setcaps):
* ext/wavpack/gstwavpackenc.c: (gst_wavpack_enc_sink_set_caps):
* gst/qtdemux/qtdemux.c: (gst_qtdemux_chain), (qtdemux_parse),
(qtdemux_parse_trak):
* gst/spectrum/gstspectrum.c: (gst_spectrum_transform_ip):
Another batch of printf format fixes.

+++ ext/dts/gstdtsdec.c
@@ -475,7 +475,7 @@ gst_dtsdec_chain (GstPad * pad, GstBuffer * buf)
 {
   GstDtsDec *dts;
   guint8 *data;
-  gint64 size;
+  gint size;
   gint length, flags, sample_rate, bit_rate, frame_length;
   GstFlowReturn result = GST_FLOW_OK;
